[bode] Add layout as hiera override layer
Add a layout specific layer. This is a requirement
for allowing the creation of multiple layouts
to deploy the same application in different ways.

When you are choosing to make certain deployment decisions,
you may want to customize the deployment's data based on how
roles are combined and apply those changes to those groups
of roles as opposed to just to a specific role.

This patch allows those things to be overridden in hiera.